refactor: 캐시 구조 개선 - CacheTemplate 인터페이스 도입 및 CachePolicy 분리(DIP 적용)
CacheTemplate 인터페이스 추가:
- RedisCacheTemplate이 CacheTemplate 인터페이스 구현
- 테스트용 NoOpCacheTemplate 추가 (캐시 미사용)
- DIP 적용으로 구현체 교체 용이

CachePolicy를 Domain 레이어로 이동:
- CacheKeyType(redis 모듈) → CachePolicy(domain/cache)
- 도메인 지식(캐시 정책)이 올바른 레이어에 위치

CacheKeys 유틸리티 클래스 제거:
- SimpleCacheKey.of() 팩토리 메서드 추가
- CachePolicy와 SimpleCacheKey 직접 조합 방식으로 단순화

Repository/Facade 의존성 변경:
- RedisCacheTemplate → CacheTemplate 인터페이스 의존
- CacheKeys 유틸리티 → SimpleCacheKey.of() + CachePolicy 직접 사용
