+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/brand/BrandRepositoryImpl.java
@@ -1,9 +1,10 @@
 package com.loopers.infrastructure.brand;
 
-import com.loopers.cache.RedisCacheTemplate;
+import com.loopers.cache.CacheTemplate;
+import com.loopers.cache.SimpleCacheKey;
 import com.loopers.domain.brand.Brand;
+import com.loopers.domain.cache.CachePolicy;
 import com.loopers.domain.brand.BrandRepository;
-import com.loopers.infrastructure.cache.CacheKeys;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -15,12 +16,12 @@ import java.util.Optional;
 public class BrandRepositoryImpl implements BrandRepository {
 
   private final BrandJpaRepository jpaRepository;
-  private final RedisCacheTemplate cacheTemplate;
+  private final CacheTemplate cacheTemplate;
 
   @Override
   public Optional<Brand> findById(Long id) {
     Brand brand = cacheTemplate.getOrLoad(
-        CacheKeys.brand(id),
+        SimpleCacheKey.of(CachePolicy.BRAND.buildKey(id), CachePolicy.BRAND.getTtl(), Brand.class),
         () -> jpaRepository.findById(id).orElse(null)
     );
     return Optional.ofNullable(brand);
